renamed LoginController
